Block transactions if user is not whitelisted
